organize inputs
