+++ apps/api/v2/src/modules/endpoints/bookings/booking.repository.ts
@@ -1,4 +1,4 @@
-import { CreateBookingInput } from "@/modules/endpoints/bookings/inputs/create-booking";
+import { CreateBookingInput } from "@/modules/endpoints/bookings/inputs/create-booking.input";
 import { PrismaReadService } from "@/modules/services/prisma/prisma-read.service";
 import { PrismaWriteService } from "@/modules/services/prisma/prisma-write.service";
 import { Injectable } from "@nestjs/common";
