+++ apps/api/v2/src/modules/endpoints/bookings/inputs/create-booking.input.ts
