+++ apps/api/v2/src/modules/endpoints/oauth-clients/inputs/update-oauth-client.input.ts
